backend/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
